refactor: extract generator to _internal module
