+++ src/conformly/_internal/generator/types/email.py
@@ -1,8 +1,9 @@
 import string
 
-from ..._internal.types import MAX_GENERATION_ATTEMPTS, ViolationType
-from ...resolver.semantics import StringSemantic
 from ..context import GenerationContext
+
+from conformly._internal.types import MAX_GENERATION_ATTEMPTS, ViolationType
+from conformly.resolver.semantics import StringSemantic
 
 LOCAL_CHARS = string.ascii_lowercase + string.digits + "._%+-"
