+++ src/conformly/_internal/generator/types/float.py
@@ -1,8 +1,9 @@
 import math
 
-from ..._internal.types import FLOAT_MAX, FLOAT_MIN, ViolationType
-from ...resolver.semantics import NumericSemantic
 from ..context import GenerationContext
+
+from conformly._internal.types import FLOAT_MAX, FLOAT_MIN, ViolationType
+from conformly.resolver.semantics import NumericSemantic
 
 
 def generate_value(
