+++ src/conformly/_internal/generator/types/uuid.py
@@ -1,9 +1,10 @@
 import string
 import uuid
 
-from ..._internal.types import ViolationType
-from ...resolver.semantics import UUIDSemantic
 from ..context import GenerationContext
+
+from conformly._internal.types import ViolationType
+from conformly.resolver.semantics import UUIDSemantic
 
 HEX_CHARS = string.hexdigits.lower()
 INVALID_CHARS = "!@#$%^&*()ghijklmnopqrstuvwxyz"
