+++ src/conformly/_internal/planner/session.py
@@ -1,5 +1,5 @@
+from conformly._internal.generator import GenerationContext
 from conformly._internal.types import CasesStrategy, FieldPath
-from conformly.generator import GenerationContext
 from conformly.resolver import ResolvedModel
 
 NameIndexMap = tuple[tuple[FieldPath, str], ...]
