+++ tests/test_generator/test_generator_orchestration.py
@@ -1,5 +1,11 @@
 import pytest
 
+from conformly._internal.generator.context import GenerationContext
+from conformly._internal.generator.orchestration import (
+    generate_field,
+    generate_invalid,
+    generate_valid,
+)
 from conformly._internal.parser import FieldSpec
 from conformly._internal.planner import PlannedTask
 from conformly._internal.types import (
@@ -10,12 +16,6 @@ from conformly._internal.types import (
     LengthRange,
     Range,
     ViolationType,
-)
-from conformly.generator.context import GenerationContext
-from conformly.generator.orchestration import (
-    generate_field,
-    generate_invalid,
-    generate_valid,
 )
 from conformly.resolver import ResolvedField, ResolvedModel
 from conformly.resolver.semantics import (
