+++ tests/test_generator/test_types_generators/test_enum_generator.py
@@ -1,8 +1,11 @@
 import pytest
 
+from conformly._internal.generator.context import GenerationContext
+from conformly._internal.generator.types.enum import (
+    _generate_not_allowed_value,
+    generate_value,
+)
 from conformly._internal.types import ViolationType
-from conformly.generator.context import GenerationContext
-from conformly.generator.types.enum import _generate_not_allowed_value, generate_value
 from conformly.resolver.semantics import EnumSemantic
 
 # ===== TESTS for generate_value() =====
