+++ tests/test_generator/test_types_generators/test_ipany_generator.py
@@ -3,9 +3,9 @@ import ipaddress
 
 import pytest
 
+from conformly._internal.generator.context import GenerationContext
+from conformly._internal.generator.types.ipvany import generate_value
 from conformly._internal.types import FieldKind, LengthRange, ViolationType
-from conformly.generator.context import GenerationContext
-from conformly.generator.types.ipvany import generate_value
 from conformly.resolver.semantics import StringSemantic
