+++ tests/test_generator/test_types_generators/test_ipv4_generator.py
@@ -1,13 +1,13 @@
 import pytest
 
-from conformly._internal.types import FieldKind, LengthRange, ViolationType
-from conformly.generator.context import GenerationContext
-from conformly.generator.types.ipv4 import (
+from conformly._internal.generator.context import GenerationContext
+from conformly._internal.generator.types.ipv4 import (
     INVALID_FORMATS,
     _generate_invalid_ipv4,
     _generate_valid_ipv4,
     generate_value,
 )
+from conformly._internal.types import FieldKind, LengthRange, ViolationType
 from conformly.resolver.semantics import StringSemantic
