+++ tests/test_generator/test_types_generators/test_ipv6_generator.py
@@ -2,9 +2,8 @@ import ipaddress
 
 import pytest
 
-from conformly._internal.types import FieldKind, LengthRange, ViolationType
-from conformly.generator.context import GenerationContext
-from conformly.generator.types.ipv6 import (
+from conformly._internal.generator.context import GenerationContext
+from conformly._internal.generator.types.ipv6 import (
     INVALID_FORMATS,
     TEST_PREFIXES,
     _compress_ipv6,
@@ -12,6 +11,7 @@ from conformly.generator.types.ipv6 import (
     _generate_valid_ipv6,
     generate_value,
 )
+from conformly._internal.types import FieldKind, LengthRange, ViolationType
 from conformly.resolver.semantics import StringSemantic
