+++ tests/test_generator/test_types_generators/test_list_generator.py
@@ -3,8 +3,8 @@ from unittest.mock import Mock, patch
 
 import pytest
 
+from conformly._internal.generator.types.list import generate_value
 from conformly._internal.types import UNSET, ViolationType
-from conformly.generator.types.list import generate_value
 from conformly.resolver.semantics import ListSemantic
 
 if TYPE_CHECKING:
@@ -28,7 +28,8 @@ def test_generates_list_of_length_1_to_3(ctx, list_sem) -> None:
     for expected_len in [1, 2, 3]:
         ctx.rng.randint = Mock(return_value=expected_len)
         with patch(
-            "conformly.generator.orchestration.generate_field", return_value="v"
+            "conformly._internal.generator.orchestration.generate_field",
+            return_value="v",
         ):
             result = generate_value(ctx, list_sem, None)
             assert len(result) == expected_len
@@ -38,7 +39,7 @@ def test_delegates_generation_to_orchestrator(ctx, mock_elem_semantic) -> None:
     ctx.rng.randint = Mock(return_value=3)
 
     with patch(
-        "conformly.generator.orchestration.generate_field", return_value="x"
+        "conformly._internal.generator.orchestration.generate_field", return_value="x"
     ) as mock_gen:
         result = generate_value(
             ctx, ListSemantic(element_semantic=mock_elem_semantic), None
@@ -58,7 +59,7 @@ def test_violation_applied_to_single_random_element(ctx, list_sem) -> None:
     ctx.rng.randint = Mock(side_effect=[3, 1])
 
     with patch(
-        "conformly.generator.orchestration.generate_field", return_value="v"
+        "conformly._internal.generator.orchestration.generate_field", return_value="v"
     ) as mock_gen:
         generate_value(ctx, list_sem, ViolationType.TOO_SHORT)
 
@@ -73,7 +74,9 @@ def test_violation_applied_to_single_random_element(ctx, list_sem) -> None:
 def test_missing_field_fallback_returns_valid_element(ctx, list_sem) -> None:
     ctx.rng.randint = Mock(side_effect=[2, 0])
 
-    with patch("conformly.generator.orchestration.generate_field") as mock_gen:
+    with patch(
+        "conformly._internal.generator.orchestration.generate_field"
+    ) as mock_gen:
         mock_gen.side_effect = ["valid", "valid2"]
 
         result = generate_value(ctx, list_sem, ViolationType.MISSING_FIELD)
@@ -91,7 +94,7 @@ def test_nested_model_propagated_to_synthetic_field(ctx, mock_elem_semantic) -> 
     ctx.rng.randint = Mock(return_value=1)
 
     with patch(
-        "conformly.generator.orchestration.generate_field", return_value={}
+        "conformly._internal.generator.orchestration.generate_field", return_value={}
     ) as mock_gen:
         generate_value(ctx, semantic, None)
 
@@ -104,7 +107,7 @@ def test_no_violation_passed_when_none(ctx, list_sem) -> None:
     ctx.rng.randint = Mock(return_value=2)
 
     with patch(
-        "conformly.generator.orchestration.generate_field", return_value="v"
+        "conformly._internal.generator.orchestration.generate_field", return_value="v"
     ) as mock_gen:
         generate_value(ctx, list_sem, violation=None)
